==> implementations/gee_zonal/country_dataset/1_Run Zonal Stats for Multiple Datasets.ipynb <==
This first notebook uses the ZonalStats python class to submit several tasks to the Google Earth Engine API. Each task requests country-level zonal stats from a specifc dataset. This results in csv tables that then need to be merged into one dataset.

import sys
sys.path.append('../../src') # relative path where the library is stored
# alternatively sys.path.append('C/Users/wb514197/Repos/GEE_Zonal/src')

from gee_tools import Catalog, ZonalStats, ZonalStats_Tsinghua, ZonalStats_Modis
import ee
import pandas as pd
import os
import glob

ee.Initialize()

cat = Catalog()

## NDVI

results = cat.search_tags("ndvi")

results = results.search_title("landsat")

results = results.search_title("8-day")

results.datasets

# AOIs = ee.FeatureCollection('users/afche18/Admin0_Polys_1000m')
AOIs = ee.FeatureCollection('users/afche18/Ethiopia_AOI') # for testing

# we try to run a separate task for each temporal statistic
temp_stats = ["mean", "max", "min", "stddev"]

# dictionary to collect and monitor tasks
ndvi_zs = {}

?ZonalStats

for landsat_collection in results.datasets.id:
    print(f"Processing {landsat_collection}")
    lc_id = landsat_collection[landsat_collection.find("/")+1:landsat_collection.find("/")+5]
    
    collection_results = []
    for stat in temp_stats:

        zs = ZonalStats(
            collection_id=landsat_collection, 
            target_features=AOIs,
            statistic_type="all", 
            output_name=f"{lc_id}_ndvi_{stat}",
            scale=1000,
            min_threshold=0.1,
            water_mask=True,
            frequency='annual',
            temporal_stat=stat,
            tile_scale = 16,
            start_year = 2020,
            end=year 2022
        )
        zs.runZonalStats()
        collection_results.append(zs)

    ndvi_zs[lc_id] = collection_results

for stat in temp_stats:

    zs = ZonalStats(
        collection_id="LANDSAT/LC08/C01/T1_8DAY_NDVI", 
        target_features=AOIs,
        statistic_type="all", 
        output_name=f"L8_ndvi_{stat}",
        scale=1000,
        min_threshold=0.1,
        water_mask=True,
        frequency='annual',
        temporal_stat=stat,
        tile_scale = 16,
        start_year = 2018,
        end_year = 2022
    )
    zs.runZonalStats()
    collection_results.append(zs)


for lc_id, zs_list in ndvi_zs.items():
    print(lc_id)
    for zs in zs_list:
        zs.reportRunTime()
#         print("\t"+zs.task.status()['state'])

## EVI

results = cat.search_tags("evi")
results = results.search_title("landsat")
results = results.search_title("8-day")

results.datasets

evi_zs = {}

for landsat_collection in results.datasets.id:
    print(f"Processing {landsat_collection}")
    lc_id = landsat_collection[landsat_collection.find("/")+1:landsat_collection.find("/")+5]
    
    collection_results = []
    for stat in temp_stats:

        zs = ZonalStats(
            collection_id=landsat_collection, 
            target_features=AOIs,
            statistic_type="all", 
            output_name=f"{lc_id}_evi_{stat}",
            scale=2000,
            min_threshold=0.1,
            water_mask=True,
            frequency='annual',
            temporal_stat=stat,
            tile_scale = 16,
        )
        zs.runZonalStats()
        collection_results.append(zs)

    evi_zs[lc_id] = collection_results

for lc_id, zs_list in evi_zs.items():
    print(lc_id)
    for zs in zs_list:
        zs.reportRunTime()
#         zs.task.cancel()
#         print("\t"+zs.task.status()['state'])

## NDSI

results = cat.search_tags("ndsi")
results = results.search_title("landsat")
results = results.search_title("8-day")

results.datasets

ndsi_zs = {}

for landsat_collection in results.datasets.id:
    print(f"Processing {landsat_collection}")
    lc_id = landsat_collection[landsat_collection.find("/")+1:landsat_collection.find("/")+5]

    collection_results = []
    for stat in temp_stats:

        zs = ZonalStats(
            collection_id=landsat_collection, 
            target_features=AOIs,
            statistic_type="all", 
            output_name=f"{lc_id}_ndsi_{stat}",
            scale=1000,
            frequency='annual',
            temporal_stat=stat,
            tile_scale = 16,
            start_year = 2021,
            end_year = 2022
        )
        zs.runZonalStats()
        collection_results.append(zs)

    ndsi_zs[lc_id] = collection_results

for lc_id, zs_list in ndsi_zs.items():
    print(lc_id)
    for zs in zs_list:
        zs.reportRunTime()
#         zs.task.cancel()
#         print("\t"+zs.task.status()['state'])

## NDWI

results = cat.search_tags("ndwi")
results = results.search_title("landsat")
results = results.search_title("8-day")

results.datasets

ndwi_zs = {}

for landsat_collection in results.datasets.id:
    print(f"Processing {landsat_collection}")
    lc_id = landsat_collection[landsat_collection.find("/")+1:landsat_collection.find("/")+5]

    collection_results = []
    for stat in temp_stats:

        zs = ZonalStats(
            collection_id=landsat_collection, 
            target_features=AOIs,
            statistic_type="all", 
            output_name=f"{lc_id}_ndwi_{stat}",
            scale=10000,
            frequency='annual',
            temporal_stat=stat,
            tile_scale = 16,
            start_year = 2021,
            end_year = 2022
        )
        zs.runZonalStats()
        collection_results.append(zs)

    ndsi_zs[lc_id] = collection_results

## LST

cat_res = cat.search_tags("temperature")
cat_res = cat_res.search_title("ECMWF")

cat_res.datasets

col_id_lst = 'ECMWF/ERA5_LAND/MONTHLY'

results_temp = []

for stat in temp_stats:
    zs = ZonalStats(
        collection_id=col_id_lst,
        band = "temperature_2m",
        target_features=AOIs,
        statistic_type="all", 
        output_name=f"lst_{stat}",
        scale=10000,
        frequency='annual',
        temporal_stat=stat,
        tile_scale = 16,
    )
    zs.runZonalStats()
    results_temp.append(zs)

for zs in results_temp:
    zs.reportRunTime()

## CHIRPS

cat_res = cat.search_title("CHIRPS")

cat_res.datasets

col_id_rain = 'UCSB-CHG/CHIRPS/PENTAD'

results_chirps = []

zs_all = ZonalStats(
    collection_id=col_id_rain,
    band = "precipitation",
    target_features=AOIs,
    statistic_type="all", 
    output_name=f"chirps",
    scale=1000,
    frequency='annual',
    temporal_stat='sum',
    tile_scale = 16,
)
zs_all.runZonalStats()

zs_sum = ZonalStats(
    collection_id=col_id_rain,
    band = "precipitation",
    target_features=AOIs,
    statistic_type="sum", 
    output_name=f"chirps_sum",
    scale=1000,
    frequency='annual',
    temporal_stat='sum',
    tile_scale = 16
)
zs.runZonalStats()

zs_all.reportRunTime()

zs_sum.reportRunTime()

## Night Lights

temp_stats = ['sum', 'mean', 'max']

dmps = 'NOAA/DMSP-OLS/NIGHTTIME_LIGHTS'

results_dmps = []

for stat in temp_stats:
    zs = ZonalStats(
        collection_id=dmps,
        band = "stable_lights",
        target_features=AOIs,
        statistic_type="all", 
        output_name=f"dmps_{stat}",
        scale=1000,
        frequency='annual',
        temporal_stat=stat,
        tile_scale = 16,
        min_threshold = 0
    )
    zs.runZonalStats()
    results_dmps.append(zs)

viirs = 'NOAA/VIIRS/DNB/MONTHLY_V1/VCMSLCFG'

for stat in temp_stats:
    zs = ZonalStats(
        collection_id=viirs,
        band = "avg_rad",
        target_features=AOIs,
        statistic_type="all", 
        output_name=f"viirs_{stat}",
        scale=1000,
        frequency='annual',
        temporal_stat=stat,
        tile_scale = 16,
        min_threshold = 0
    )
    zs.runZonalStats()
    results_viirs.append(zs)

## Impervious

zs = ZonalStats_Tsinghua(
    target_features=AOIs,
    output_name=f"impervious",
    scale=10000,
    tile_scale = 16
)
zs.runZonalStats()

zs.reportRunTime()

## Cropland

zs_crop = ZonalStats_Modis(
    target_features=AOIs,
    output_name=f"cropland",
    scale=10000,
    tile_scale = 16
)
zs_crop.runZonalStats()

zs_crop.reportRunTime()
