==> implementations/gee_zonal/country_dataset/3_Merge Data.ipynb <==
This third notebook collects datasets from step 2, and merges them into one comprehensive table.

import sys

sys.path.append("../../src")  # relative path where the library is stored

import pandas as pd
import os

data_dir = os.path.join("X:/data/ee")
out_dir = os.path.join(data_dir, "output")

# pd.set_option('display.max_rows', None)

# pd.set_option('display.max_columns', None)

ndvi = pd.read_csv(
    os.path.join(out_dir, "ndvi.csv"), encoding="UTF-8"
)  # index_col=[0,1]
ndvi.set_index(["wb_adm0_na", "year"], inplace=True)
ndvi = ndvi.loc[ndvi.landsat_id != "LT04"].copy()
ndvi.loc[:, "landsat_id"] = pd.Categorical(
    ndvi["landsat_id"], categories=["LE07", "LT05", "LC08"], ordered=True
)
ndvi.sort_values(["wb_adm0_na", "year", "landsat_id"], inplace=True)
ndvi.reset_index(inplace=True)
ndvi.drop_duplicates(["wb_adm0_na", "year"], keep="first", inplace=True)
ndvi.set_index(["wb_adm0_na", "year"], inplace=True)
ndvi = ndvi.loc[:, ndvi.columns[["ndvi" in col for col in ndvi.columns]]].copy()

evi = pd.read_csv(os.path.join(out_dir, "evi.csv"), encoding="UTF-8")
evi.set_index(["wb_adm0_na", "year"], inplace=True)
evi = evi.loc[evi.landsat_id != "LT04"].copy()
evi.loc[:, "landsat_id"] = pd.Categorical(
    evi["landsat_id"], categories=["LE07", "LT05", "LC08"], ordered=True
)
evi.sort_values(["wb_adm0_na", "year", "landsat_id"], inplace=True)
evi.reset_index(inplace=True)
evi.drop_duplicates(["wb_adm0_na", "year"], keep="first", inplace=True)
evi.set_index(["wb_adm0_na", "year"], inplace=True)
evi = evi.loc[:, evi.columns[["evi" in col for col in evi.columns]]].copy()

ndsi = pd.read_csv(os.path.join(out_dir, "ndsi.csv"), encoding="UTF-8")
ndsi.set_index(["wb_adm0_na", "year"], inplace=True)
ndsi = ndsi.loc[ndsi.landsat_id != "LT04"].copy()
ndsi.loc[:, "landsat_id"] = pd.Categorical(
    ndsi["landsat_id"], categories=["LE07", "LT05", "LC08"], ordered=True
)
ndsi.sort_values(["wb_adm0_na", "year", "landsat_id"], inplace=True)
ndsi.reset_index(inplace=True)
ndsi.drop_duplicates(["wb_adm0_na", "year"], keep="first", inplace=True)
ndsi.set_index(["wb_adm0_na", "year"], inplace=True)
ndsi = ndsi.loc[:, ndsi.columns[["ndsi" in col for col in ndsi.columns]]].copy()

ndwi = pd.read_csv(os.path.join(out_dir, "ndwi.csv"), encoding="UTF-8")
ndwi.set_index(["wb_adm0_na", "year"], inplace=True)
ndwi = ndwi.loc[ndwi.landsat_id != "LT04"].copy()
ndwi.loc[:, "landsat_id"] = pd.Categorical(
    ndwi["landsat_id"], categories=["LE07", "LT05", "LC08"], ordered=True
)
ndwi.sort_values(["wb_adm0_na", "year", "landsat_id"], inplace=True)
ndwi.reset_index(inplace=True)
ndwi.drop_duplicates(["wb_adm0_na", "year"], keep="first", inplace=True)
ndwi.set_index(["wb_adm0_na", "year"], inplace=True)
ndwi = ndwi.loc[:, ndwi.columns[["ndwi" in col for col in ndwi.columns]]].copy()

chirps = pd.read_csv(
    os.path.join(out_dir, "chirps.csv"), encoding="UTF-8"
)  # index_col=[0,1]
chirps.set_index(["wb_adm0_na", "year"], inplace=True)
chirps.sort_values(["wb_adm0_na", "year"], inplace=True)
chirps = chirps.loc[
    :, chirps.columns[["precipitation" in col for col in chirps.columns]]
].copy()

lst = pd.read_csv(os.path.join(out_dir, "temperature.csv"), encoding="UTF-8")
lst.set_index(["wb_adm0_na", "year"], inplace=True)
lst.sort_values(["wb_adm0_na", "year"], inplace=True)
lst = lst.loc[:, lst.columns[["temperature" in col for col in lst.columns]]].copy()

dmps = pd.read_csv(os.path.join(out_dir, "dmps.csv"), encoding="UTF-8")
dmps.set_index(["wb_adm0_na", "year"], inplace=True)
dmps = dmps.loc[:, dmps.columns[["lights" in col for col in dmps.columns]]].copy()
dmps.loc[:, "lights_source"] = "DMPS"

viirs = pd.read_csv(os.path.join(out_dir, "viirs.csv"), encoding="UTF-8")
viirs.set_index(["wb_adm0_na", "year"], inplace=True)
viirs = viirs.loc[:, viirs.columns[["lights" in col for col in viirs.columns]]].copy()
viirs.loc[:, "lights_source"] = "VIIRS"

lights = pd.concat([dmps, viirs])
lights.sort_values(["wb_adm0_na", "year"], inplace=True)

cropland = pd.read_csv(os.path.join(out_dir, "cropland.csv"), encoding="UTF-8")
cropland.set_index(["wb_adm0_na", "year"], inplace=True)
cropland.sort_values(["wb_adm0_na", "year"], inplace=True)
cropland = cropland[["cropland"]].copy()

impervious = pd.read_csv(os.path.join(out_dir, "impervious.csv"), encoding="UTF-8")
impervious.set_index(["wb_adm0_na", "year"], inplace=True)
impervious.sort_values(["wb_adm0_na", "year"], inplace=True)
impervious.rename(columns={"imperv": "impervious"}, inplace=True)
impervious = impervious[["impervious"]].copy()

full_all = (
    ndvi.join(evi, how="outer")
    .join(ndsi, how="outer")
    .join(ndwi, how="outer")
    .join(chirps, how="outer")
    .join(lst, how="outer")
    .join(lights, how="outer")
    .join(cropland, how="outer")
    .join(impervious, how="outer")
)

full_all.head()

All of the data is indexed by country name, this final step re-merges some key country attributes that weren't preserved in the earlier steps (iso code etc.)

full_all.sort_values(["wb_adm0_na", "year"], inplace=True)

adm0 = pd.read_csv(os.path.join(data_dir, "Admin0_Polys_v3.csv"), encoding="UTF-8")

adm0 = adm0.loc[adm0.WDI == 1].copy()

adm0.drop(
    ["WDI", "Data", "FID_100", "latY", "longX", "Shape_Leng", "Shape_Area"],
    axis=1,
    inplace=True,
)

adm0.rename(str.lower, axis=1, inplace=True)

full_all.reset_index(inplace=True)

full_all.rename(columns={"wb_adm0_na": "wb_adm0_na2"}, inplace=True)

full_all_adm = full_all.merge(adm0, on="wb_adm0_na2")

full_all_adm.drop("wb_adm0_na2", inplace=True, axis=1)

# full_all_adm.wb_adm0_na.unique()

valid_codes = adm0.wb_adm0_na.unique()

full_all_adm = full_all_adm.loc[full_all_adm.wb_adm0_na.isin(valid_codes)].copy()

len(full_all_adm.wb_adm0_na.unique())

cols = adm0.columns.append(full_all.columns)

cols = cols.drop("wb_adm0_na2")

full_all_adm = full_all_adm[cols[~cols.duplicated()]].copy()

full_all_adm = full_all_adm.sort_values(["wb_adm0_na", "year"])

full_all_adm.shape

len(full_all_adm.year.unique())

len(full_all_adm.wb_adm0_na.unique())

218 * 41

full_all_adm.to_csv(os.path.join(out_dir, "Country Dataset June7.csv"))
